backend/requirements.txt
celery==5.3.4
Django==4.1
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.0
redis==5.0.1
